pyproject.toml
[project]
name = "dataops"
version = "0.1.0"
description = ""
authors = [
    {name = "ebenkara15",email = "erwan.benkara@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12, <3.13"
dependencies = [
    "dvc (>=3.59.2,<4.0.0)",
    "mlflow (>=2.22.0,<3.0.0)",
    "scikit-learn (>=1.6.1,<2.0.0)",
    "evidently (>=0.7.6,<0.8.0)",
    "dbt-duckdb (>=1.9.3,<2.0.0)",
    "requests (>=2.32.3,<3.0.0)",
    "apache-airflow (==2.10.4)",
    "dbt-bigquery (==1.9.2)",
    "apache-airflow-providers-cncf-kubernetes (>=10.5.0,<11.0.0)",
]

[tool.poetry]
packages = [{include = "dataops", from = "src"}]

[tool.poetry.group.dev.dependencies]
ruff = "^0.11.12"
mypy = "^1.16.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
